Move tests to a separate module
